backend/requirements.txt
aiogram~=3.21.0
APScheduler~=3.11.0
pytest~=8.4.1
pytest-django~=4.8.0
pytest-asyncio~=0.23.0
Django~=5.2.4
djangorestframework~=3.16.0
aiohttp~=3.9.0
python-decouple~=3.8